refactor: cleaned up file directory, removed redundant imports and page routing, made naming conventions consistent
