+++ src/App.js
@@ -2,7 +2,7 @@
 import { Container } from 'react-bootstrap'
 import { AuthProvider } from "./user-auth/contexts/AuthContexts";
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
-import Landing from './Landing'
+import Landing from './user-auth/Landing'
 import PersistentDrawerLeft from './Main.tsx'
 import SignUp from "./user-auth/SignUp"
 import LogIn from "./user-auth/LogIn"
