--- src/Pages/ActivityPage.js
@@ -1,5 +1,0 @@
-import Container from "@mui/material/Container";
-export function ActivityPage() {
-    return <Container>Activity</Container>;
-}
-export default ActivityPage;
