+++ src/Pages/ActivityPage/ActivityPage.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export default function Activity() {
+export default function ActivityPage() {
     return (
         <div>Activity page yo</div>
     )
