--- src/Pages/FoodPage.js
@@ -1,5 +1,0 @@
-import Container from "@mui/material/Container";
-export function FoodPage() {
-    return <Container>Food</Container>;
-}
-export default FoodPage;
