+++ src/Pages/RESTTestButtons/REStButtonTestPage.js
@@ -1,8 +1,8 @@
 
 import Container from "@mui/material/Container";
-import GetActivitiesRequestButton from "./RESTTestButtons/GetActivitiesRequestButton";
-import GetDaysRequestButton from "./RESTTestButtons/GetDaysRequestButton";
-import GetFoodsRequestButton from "./RESTTestButtons/GetFoodsRequestButton";
+import GetActivitiesRequestButton from "./Components/GetActivitiesRequestButton";
+import GetDaysRequestButton from "./Components/GetDaysRequestButton";
+import GetFoodsRequestButton from "./Components/GetFoodsRequestButton";
 
 export function REStButtonTestPage() {
     return <Container>RESTButtonTestPage
