+++ src/user-auth/SignUp.js
@@ -23,7 +23,7 @@ export default function SignUp() {
             setError('')
             setLoading(true)
             await signup(emailRef.current.value, passwordRef.current.value)
-            navigate('/dash')
+            navigate('/DashPage')
         } catch {
             setError('Failed to create an account')
         }
